fine

+++ src/format.rs
@@ -451,7 +451,7 @@ impl RaplEnergyCounter {
             .map(|now_value| {
                 let delta_dur = Instant::now() - self.start_instant;
                 let delta_uj = now_value - self.start_value;
-                let value = delta_uj as u128 * 10u128.pow(6) / delta_dur.as_micros();
+                let value = delta_uj as f64 * 10f64.powf(6.) / delta_dur.as_micros() as f64;
                 Power::from_microwatts(value as f64)
             })
     }
